Partition the conditional variable in PointLockManager to reduce inefficient waiter wake up
Summary:
PointLockManager manages point lock per key. The old implementation
partition the per key lock into 16 stripes. Each stripe handles the
point lock for a subset of keys. Each stripe have only one conditional
variable. This conditional variable is used by all the transactions that
are waiting for its turn to acquire a lock of a key that belongs to this
stripe.

In production, we notified that when there are multiple transactions
trying to write to the same key, all of them will wait on the same
conditional variables. When the previous lock holder released the key,
all of the transactions are waked up, but only one of them could
proceed, and the rest goes back to sleep. This wasted a lot of CPU
cycles.

In order to solve this issue, we create a per key waiter queue. Each transaction
that waits for the lock will wait on a dedicated conditional variable in
the waiter queue. When previous lock holder released the lock, it wakes
up the next transaction waiting in the waiter queue if it is exclusive
lock, or multiple transactions if they are requesting shared lock. The
queue respect FIFO order, except it prioritize an exclusive lock request
if the transaction is already holding a shared lock.

Test Plan:

Unit test

Reviewers:

Subscribers:

Tasks:

Tags: